src/playground/requirements.txt
python-socketio
fastapi
uvicorn
python-dotenv
requests
kubernetes